ux/views: factored {product} out of the {native}, {isce2}, and {nisar} readers in {explorer}
