+++ ux/client/views/explorer/reader/product.js
@@ -1,0 +1,33 @@
+// -*- web -*-
+//
+// michael a.g. aïvázis <michael.aivazis@para-sim.com>
+// (c) 1998-2024 all rights reserved
+
+
+// external
+import React from 'react'
+
+// local
+// components
+import { enumValue } from '../form'
+
+// the product type field
+export const Product = ({ name, rep, current, select }) => {
+    // assemble the behaviors
+    const behaviors = {
+        onClick: select(rep)
+    }
+    // deduce my state
+    const state = current === rep ? "selected" : "enabled"
+    // pick a field selector based on my state
+    const Selector = enumValue(state)
+    // and render it
+    return (
+        <Selector {...behaviors}>
+            {name}
+        </Selector>
+    )
+}
+
+
+// end of file
